Update functional tests- explicit port, use config, move mock-postcode
* Have an explicit port number to use as part of the call to the mock api
* Use a config for the port
* Move mock-postcode
