+++ functional-tests/apps/default.js
@@ -9,7 +9,7 @@ module.exports = {
         addressKey: 'address-one',
         apiSettings: {
           authorization: process.env.POSTCODE_AUTH,
-          hostname: 'http://localhost:8080/api/postcode-test'
+          hostname: 'http://localhost:8081/api/postcode-test'
         },
         validate: {
           allowedCountries: ['England']
