+++ functional-tests/config.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+  port: 8081
+};
