+++ functional-tests/lib/mock-postcode.js
@@ -11,7 +11,7 @@ module.exports = router.use('/api/postcode-test/:action?/:postcode?', (req, res)
   if (action === 'addresses') {
     if (req.query.postcode === 'CR0 2EU') {
       // eslint-disable-next-line camelcase
-      res.send(JSON.stringify([{formatted_address: '49 Sydenham Road\nCroydon\nCR0 2EU', postcode: 'CR0 2EU'}]));
+      res.send(JSON.stringify([{ formatted_address: '49 Sydenham Road\nCroydon\nCR0 2EU', postcode: 'CR0 2EU' }]));
     } else {
       res.setHeader('Content-Type', 'application/json');
       res.send(JSON.stringify([]));
